requirements.txt
pycryptodome==3.4
Flask==0.11
Jinja2==2.8
redis==2.10.5
captcha==0.2.1